Feat: Change the stylesheet using flex properties
Having the seperate sections to work with helped
with using flex to align them on the page more
easily. I initially used 'flex: 1 1 auto' on the
left side for the flex to take up the rest of
the unused space in the header, pushing the
'.right' side div to its furthest point within the
flex. I also set the '.right' with the value
'flex: 0 1 auto' to ensure that it doesn't try to
grow to fill the space too.

This wasn't needed. I instead could just use
'justify-content: space between' this means all
items within that flex would be spaced with empty
space inbetween instead, this is a better method.

Align-items was used to ensure that the items
would be aligned on the cross axis, in this case
the center of the flex container.