Feat: Order the webpage to center items using flex
Edited the HTML to wrap sections of content in a
way that makes sense for the want of centering
certain items together e.g. the text and image
pairs. This also allowed it to be a flex
container that I could arrange vertically and edit
that way. The main change was 'justify-content:
center' this just meant that each section of
elements would be centered on the page depending
on the size.