Feat: Rearrange the HTML and styling to match
Changed much of the ordering of the HTML. Added
divs section by section with utilising flex
containers to control the positioning elements to
match the example. Certain things had to be
aligned to one another, such as the icons and
the header. This was achieved by viewing them as
being flex container and working out in which
order to have those flex container be.